requirements.txt
google-genai
python-dotenv
aiohttp
